+++ api/QLThuVien.Business/ViewModels/BookVm.cs
@@ -11,5 +11,5 @@ public class BookVm
     public required string CategoryName { get; set; }
     public required int Count { get; set; }
     public required string ImageUrl { get; set; }
-    public float? AverageRating { get; set; }
+    public double? AverageRating { get; set; }
 }
